Move widgets specific functions to helpers.ts
