+++ widgets/src/lib/utils/ViewUtils.ts
@@ -1,5 +1,3 @@
-import type { TableData } from "../components/InferenceWidget/shared/types";
-
 const ESCAPED = {
 	'"': "&quot;",
 	"'": "&#39;",
@@ -88,42 +86,6 @@ export function scrollToMax(elt: HTMLElement, axis: "x" | "y" = "y") {
 		left: axis === "x" ? elt.scrollWidth : undefined,
 		top: axis === "y" ? elt.scrollHeight : undefined,
 	});
-}
-
-/*
-* Converts table from [[Header0, Header1, Header2], [Column0Val0, Column1Val0, Column2Val0], ...]
-* to {Header0: [ColumnVal0, ...], Header1: [Column1Val0, ...], Header2: [Column2Val0, ...]}
-*/
-export function convertTableToData(table: (string | number)[][]): TableData {
-	return Object.fromEntries(
-		table[0].map((cell, x) => {
-			return [
-				cell,
-				table
-					.slice(1)
-					.flat()
-					.filter((_, i) => i % table[0].length === x)
-					.map((x) => String(x)), // some models can only handle strings (no numbers)
-			];
-		})
-	);
-}
-
-/*
-* Converts data from {Header0: [ColumnVal0, ...], Header1: [Column1Val0, ...], Header2: [Column2Val0, ...]}
-* to [[Header0, Header1, Header2], [Column0Val0, Column1Val0, Column2Val0], ...]
-*/
-export function convertDataToTable(data: TableData): (string | number)[][] {
-	const dataArray = Object.entries(data); // [header, cell[]][]
-	const nbCols = dataArray.length;
-	const nbRows = (dataArray[0]?.[1]?.length ?? 0) + 1;
-	return Array(nbRows)
-		.fill("")
-		.map((_, y) =>
-			Array(nbCols)
-				.fill("")
-				.map((_, x) => (y === 0 ? dataArray[x][0] : dataArray[x][1][y - 1]))
-		);
 }
 
 /*
